Add Sidebar
Added sidebar to accomodate list of gateways.
