+++ assets/css/app.css
@@ -1,4 +1,5 @@
 /* This file is for your ma in application css. */
+@import "https://fonts.googleapis.com/css?family=Poppins:300,400,500,600,700";
 
 * {
   box-sizing: border-box;
@@ -62,7 +63,7 @@ body {
   -ms-flex-direction: column;
   flex-direction: column;
   font-size: 62.5%;
-  font-family: Roboto, Open Sans, Segoe UI, sans-serif;
+  font-family:'Poppins', Roboto, Open Sans, Segoe UI, sans-serif;
   font-weight: 400;
   font-style: normal;
   -webkit-text-size-adjust: 100%;
@@ -71,6 +72,7 @@ body {
   text-rendering: optimizeLegibility;
   font-feature-settings: "pnum";
   font-variant-numeric: proportional-nums;
+  margin-top: 0 !important;
 }
 
 .globalContent {
@@ -816,3 +818,77 @@ footer {
   font-size: 30px;
   font-weight: bold;
 }
+
+.wrapper {
+  display: flex;
+  align-items: stretch;
+}
+
+a, a:hover, a:focus {
+  color: inherit;
+  text-decoration: none;
+  transition: all 0.3s;
+}
+
+#sidebar {
+  min-width: 250px;
+  max-width: 250px;
+  min-height: 100vh;
+  background: #9169d8;
+  color: #fff;
+  transition: all 0.3s;
+}
+
+#sidebar.active {
+  margin-left: -250px;
+}
+
+@media (max-width: 768px) {
+  #sidebar {
+      margin-left: -250px;
+  }
+  #sidebar.active {
+      margin-left: 0;
+  }
+}
+
+a, a:hover, a:focus {
+  color: inherit;
+  text-decoration: none;
+  transition: all 0.3s;
+}
+
+#sidebar .sidebar-header {
+  padding: 20px;
+  background: #9169d8;
+}
+
+#sidebar ul.components {
+  padding: 20px 0;
+  border-bottom: 1px solid #47748b;
+}
+
+#sidebar ul p {
+  color: #fff;
+  padding: 10px;
+}
+
+#sidebar ul li a {
+  padding: 10px;
+  font-size: 1.1em;
+  display: block;
+  color: #fff;
+}
+
+#sidebar ul li a:hover {
+  color: #9169d8;
+  background: #fff;
+}
+
+ul ul a {
+  font-size: 0.9em !important;
+  padding-left: 30px !important;
+  background: hsl(229, 48%, 61%);
+}
+
+
