+++ assets/js/sidebar.js
@@ -1,0 +1,10 @@
+
+export function sidebar(){
+  console.log('test sidebar');
+}
+
+export function sidebarCheck() {
+  document.getElementById('sidebar').onclick = function(){
+    console.log('sidebar check called');
+  }
+}
