+++ lib/gringotts_payment_web/templates/layout/sidebar.html.eex
@@ -1,0 +1,9 @@
+<div class="sidebar-header">
+  <h3>Gateways</h3>
+</div>
+
+<ul class="list-unstyled components">
+  <li class="active"><%= link "Stripe", to: page_path(@conn, :index) %></li>
+  <li><a href="#">AuthorizeNet</a></li>
+  <li><a href="#">Contact</a></li>
+</ul>
